# Faster reverse complement using translation table